Move couchbase custom transcoder code to transcoder.go
